ponowny reformat kodu
