+++ src/common/dataset.py
@@ -1,6 +1,5 @@
 from src.common.read_txt import ReadTXT
 from src.common.save_txt import SaveTXT
-import PySimpleGUI as sg
 from typing import Dict, List, Tuple
 from src.common.static_method import get_columns_headers, get_file_path
 import os
@@ -52,37 +51,8 @@ class Dataset:
         else:
             return file_path
 
-    def get_console_layout(self) -> str:
-        column_max_width = self.calculate_column_char_width()
-        header_layout = "|"
-        for column_name in ["Lp."] + self.__columns_headers:
-            header_layout += "{:^{}}|".format(
-                column_name, column_max_width[column_name]
-            )
-        data_layout = ""
-        for index, row in self.__data.items():
-            data_layout += "|{:^{}}|".format(str(index), column_max_width["Lp."])
-            for column_name in self.__columns_headers:
-                data_layout += "{:^{}}|".format(
-                    row[column_name], column_max_width[column_name]
-                )
-            data_layout += "\n"
-        tab_size = len(header_layout)
-        console_layout = "{:-^{}}".format("", tab_size)+"\n"+header_layout+"\n"+"{:-^{}}".format("", tab_size)+"\n"+data_layout+"{:-^{}}".format("", tab_size)
-        return console_layout
-
-    def get_gui_layout(self):
-        column_max_width = self.calculate_column_char_width()
-        header_layout = []
-        for column_name in ["Lp."] + self.__columns_headers:
-            header_layout.append(
-                sg.In(column_name, size=(column_max_width[column_name], 1), pad=(1, 1), justification="center", do_not_clear=True, disabled=True)
-            )
-        data_layout = [header_layout]
-        for index, row in self.__data.items():
-            row_layout = []
-            row_layout.append(sg.In(f"{index}", size=(column_max_width["Lp."], 1), pad=(1, 1), justification="center", do_not_clear=True, disabled=True))
-            for column_name in self.__columns_headers:
-                row_layout.append(sg.In(row[column_name], size=(column_max_width[column_name], 1), pad=(1, 1), justification="center", key=(index, column_name), do_not_clear=True))
-            data_layout.append(row_layout)
-        return data_layout
+    def udate_data(self, values: Dict[Tuple[str,str], str]) -> None:
+        self.__data = {
+            index: {key: values[(index, key)] for key, _ in row.items()}
+            for index, row in self.__data.items()
+        }
